requirements.txt
fastapi==0.110.0
ipython==8.12.3
protobuf==5.26.1
uvicorn==0.29.0
google-generativeai
